requirements.txt
fastapi
uvicorn
aiogram
supabase
jinja2
python-dotenv
websockets==12.0